Refactor CERTIFICATE command
